requirements.txt
hydra-core
lightning
pandas
scikit-learn
torch
torchvision
umap-learn
wandb